# api/openapi.yaml
openapi: 3.0.4
info:
  title: UnderSound Content API
  description: |
    API de contenido inspirada en Bandcamp para gestionar **álbumes, pistas, merchandising, labels y comentarios**.
    Incluye paginación, filtros, subida de archivos (multipart), acciones de publicación y favoritos.
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/escuelapolitecnicaca/contenido/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.example.com/v1
  description: Producción
- url: http://localhost:8081
  description: Local
security:
- oauth2:
  - content.read
  - content.write
...
paths:
...
  /users/{userId}/favorites/albums:
    get:
      tags:
      - Favorites
      summary: Listar álbumes favoritos de un usuario
      operationId: usersUserIdFavoritesAlbumsGET
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      # Pagination removed: endpoint returns all favorites for the provided userId
      responses:
        "200":
          description: Lista paginada de álbumes favoritos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlbumList"
      x-swagger-router-controller: Favorites
    post:
      tags:
      - Favorites
      summary: Añadir álbum a favoritos
      operationId: usersUserIdFavoritesAlbumsPOST
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/favorites_albums_body"
        required: true
      responses:
        "200":
          description: OK
      x-swagger-router-controller: Favorites
  /users/{userId}/favorites/albums/{albumId}:
    delete:
      tags:
      - Favorites
      summary: Quitar álbum de favoritos
      operationId: usersUserIdFavoritesAlbumsAlbumIdDELETE
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Favorites
  /users/{userId}/favorites/merch:
    get:
      tags:
      - Favorites
      summary: Listar productos de merch favoritos de un usuario
      operationId: usersUserIdFavoritesMerchGET
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      # Pagination removed: endpoint returns all merch favorites for the provided userId
      responses:
        "200":
          description: Lista paginada de merch favoritos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMerchList"
      x-swagger-router-controller: Favorites
    post:
      tags:
      - Favorites
      summary: Añadir producto de merch a favoritos
      operationId: usersUserIdFavoritesMerchPOST
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/favorites_merch_body"
        required: true
      responses:
        "200":
          description: OK
      x-swagger-router-controller: Favorites
  /users/{userId}/favorites/merch/{merchId}:
    delete:
      tags:
      - Favorites
      summary: Quitar producto de merch de favoritos
      operationId: usersUserIdFavoritesMerchMerchIdDELETE
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: ID del usuario (acepta números o UUIDs como texto)
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Favorites
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 123
      example:
        total: 123
        limit: 20
        page: 1
    Image:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      example:
        alt: alt
        width: 6
        url: http://example.com/aeiou
        height: 1
    ArtistRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    LabelRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        artist:
          $ref: "#/components/schemas/ArtistRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        releaseDate:
          type: string
          format: date
        releaseState:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - archived
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        cover:
          $ref: "#/components/schemas/Image"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        stats:
          $ref: "#/components/schemas/Album_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseDate: 2000-01-23
        description: description
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        tracks:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        tags:
        - tags
        - tags
        cover:
          alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        createdAt: 2000-01-23T04:56:07.000+00:00
        stats:
          playCount: 9
          likeCount: 3
          ratingAvg: 4.5
          commentCount: 2
        price: 0.8008282
        genres:
        - genres
        - genres
        currency: EUR
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseState: draft
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    PaginatedAlbumList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Album"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
...
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        durationSec:
          type: integer
        trackNumber:
          type: integer
        artist:
          $ref: "#/components/schemas/ArtistRef"
        album:
          $ref: "#/components/schemas/AlbumRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        audio:
          $ref: "#/components/schemas/Track_audio"
        lyrics:
          $ref: "#/components/schemas/Track_lyrics"
        stats:
          $ref: "#/components/schemas/Track_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        trackNumber: 5
        artist: null
        stats:
          playCount: 7
        album:
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        label: null
        audio:
          codec: codec
          bitrate: 2
          url: http://example.com/aeiou
        title: title
        durationSec: 5
        lyrics:
          language: language
          text: text
        updatedAt: 2000-01-23T04:56:07.000+00:00
    AlbumRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
...
    MerchItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        label:
          $ref: "#/components/schemas/LabelRef"
        artist:
          $ref: "#/components/schemas/ArtistRef"
        availability:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        images:
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        variants:
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        availability: in_stock
        type: camiseta
        updatedAt: 2000-01-23T04:56:07.000+00:00
    Variant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        color:
          type: string
        size:
          type: string
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        stock:
          type: integer
      example:
        color: color
        size: size
        price: 0.8008282
        currency: EUR
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        sku: sku
        stock: 6
...
    PaginatedMerchList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MerchItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
...
    favorites_albums_body:
      required:
      - albumId
      type: object
      properties:
        albumId:
          type: string
          format: uuid
    favorites_merch_body:
      required:
      - merchId
      type: object
      properties:
        merchId:
          type: string
          format: uuid
    Album_stats:
      type: object
      properties:
        playCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        ratingAvg:
          type: number
          format: float
          example: 4.5
      example:
        playCount: 9
        likeCount: 3
        ratingAvg: 4.5
        commentCount: 2
...
    Track_audio:
      type: object
      properties:
        url:
          type: string
          format: uri
        bitrate:
          type: integer
        codec:
          type: string
      example:
        codec: codec
        bitrate: 2
        url: http://example.com/aeiou
    Track_lyrics:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
      example:
        language: language
        text: text
    Track_stats:
      type: object
      properties:
        playCount:
          type: integer
      example:
        playCount: 7
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.undersound.example.com/authorize
          tokenUrl: https://auth.undersound.example.com/token
          scopes:
            content.read: Lectura de contenido
            content.write: Escritura de contenido
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT